Adds first bot and Adjudicator
Include the first dummy bot that will be used to play the game.
The Adjudicator will be inside the Monopoly folder is added to the project
as a submodule. If you're syncing for the first time run the command:
git submodule update --init --recursive

To play one game:
python experiment.py

Assumptions:
You have python3 installed on your machine. AgentOne will always be player
1 and AgentTwo as player 2, although this can be changed in expreiments.py